requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
APScheduler==3.10.4
pydantic==2.9.2
google-api-python-client==2.149.0
google-auth==2.35.0
